Added Flinch to idle state
Before when you shot a idle zombie in the chest he wouldn't flinch. The zombie will now play the flinch animation when shot in the chest and then proceed to chase the player. The zombie however does not chase the player if he is shot in the limbs and the player has yet to step into its detection radius.